Sequence: Added explicit casts to appease some compilers
#jira UE-169061
#rb Henrik.Karlsson
#preflight trivial

[CL 22924780 by Andrew Rodham in ue5-main branch]

+++ Engine/Source/Runtime/MovieScene/Private/EntitySystem/MovieSceneEntityManager.cpp
@@ -987,7 +987,7 @@ void FEntityManager::EnterIteration() const
 
 void FEntityManager::ExitIteration() const
 {
-	checkSlow(IterationCount > 0);
+	checkSlow(static_cast<uint16>(IterationCount) > 0);
 	--IterationCount;
 }
